+++ src/app/(with-loading)/layout.tsx
@@ -1,0 +1,6 @@
+"use client";
+import { LoadingProvider } from "@/context/loading-context";
+
+export default function WithLoadingLayout({ children }: { children: React.ReactNode }) {
+  return <LoadingProvider>{children}</LoadingProvider>;
+}
