+++ src/app/departments/layout.tsx
@@ -1,0 +1,1 @@
+export { default } from "../(with-loading)/layout";
