+++ src/components/task/task-management.tsx
@@ -4,10 +4,9 @@ import { useState } from "react"
 import { Button } from "@/components/ui/button"
 import { Input } from "@/components/ui/input"
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs"
-import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from "@/components/ui/table"
 import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "@/components/ui/select"
-import { CircleDot, Home, Plus, RefreshCw, Search, Settings, User, Calendar as CalendarIcon, Loader2 } from "lucide-react"
-import { cn, getAvatarFromName } from "@/lib/utils"
+import { CircleDot, Home, Plus, RefreshCw, Search, Settings, User } from "lucide-react"
+import { cn } from "@/lib/utils"
 import AddTaskModal from "./add-task-modal"
 import { useCategories } from "@/hooks/userCategories"
 import { Category } from "@/types/category"
@@ -15,6 +14,7 @@ import { useTasks } from "@/hooks/useTask"
 import { Task } from "@/types/task"
 import { Tooltip } from "antd"
 import LoadingSpinner from "../ui/loading-spinner"
+import TableTask from "./table-task"
 
 export default function TaskManagement() {
   const [activeTab, setActiveTab] = useState("ongoing")
@@ -29,6 +29,8 @@ export default function TaskManagement() {
   const handleAddTask = (newTask: Task) => {
     console.log(newTask)
   }
+
+  const taskCompleted = tasks?.filter((task: Task) => task.status === "completed")
 
   return (
     <div className="mx-auto w-full">
@@ -161,85 +163,7 @@ export default function TaskManagement() {
                     <LoadingSpinner />
                   </div>
                 ) : (
-                <Table>
-                  <TableHeader>
-                    <TableRow>
-                      <TableHead className="w-10"></TableHead>
-                      <TableHead className="w-10">#</TableHead>
-                      <TableHead>Nội dung</TableHead>
-                      <TableHead className="w-24 text-center hidden md:table-cell">Trọng số</TableHead>
-                      <TableHead className="w-28 hidden md:table-cell">Hạn xử lý</TableHead>
-                      <TableHead className="w-28 hidden md:table-cell">Ngày tạo</TableHead>
-                    </TableRow>
-                  </TableHeader>
-                  <TableBody>
-                    {tasks?.map((task: Task) => (
-                      <TableRow key={task.id}>
-                        <TableCell>
-                          <input type="checkbox" className="rounded" />
-                        </TableCell>
-                        <TableCell>{task.id}</TableCell>
-                        <TableCell>
-                          <div className="flex">
-                            <div className="flex flex-col items-center justify-center min-w-[56px] max-w-[56px] mr-3">
-                              {task.mainHandler && (
-                                <Tooltip title={`Xử lý chính: ${task.mainHandler.name}`} placement="top">
-                                  <div className="w-10 h-10 rounded-full bg-blue-500 text-white flex items-center justify-center text-base font-bold cursor-default mb-1">
-                                    {getAvatarFromName(task.mainHandler.name)}
-                                  </div>
-                                </Tooltip>
-                              )}
-                            </div>
-                            <div className="flex flex-col min-w-[60px] max-w-[130px] mr-3 w-full">
-                              {task?.category && (
-                                <span className={`block bg-${task?.category?.color}-100 text-${task?.category?.color}-800 text-xs px-2 py-0.5 rounded mb-1`}>
-                                  {String(task?.category?.display_name)}
-                                </span>
-                              )}
-                              {task.assignees.length > 0 && (
-                                <div>
-                                  <span className="text-sm text-gray-500">Phối hợp: </span>
-                                  {(() => {
-                                    const MAX_DISPLAY = 3;
-                                    const displayedAssignees = task.assignees.slice(0, MAX_DISPLAY);
-                                    const hiddenCount = task.assignees.length - MAX_DISPLAY;
-                                    return (
-                                      <>
-                                        {displayedAssignees.map((assignee: any, index: number) => (
-                                          <span
-                                            title={assignee.name}
-                                            key={index}
-                                            className={
-                                              "inline-flex items-center justify-center w-6 h-6 rounded-full bg-green-100 text-green-800 text-xs " +
-                                              (index > 0 ? "-ml-2 border-2 border-white" : "")
-                                            }
-                                          >
-                                            {getAvatarFromName(assignee.name)}
-                                          </span>
-                                        ))}
-                                        {hiddenCount > 0 && (
-                                          <span className="inline-flex items-center justify-center w-6 h-6 rounded-full bg-gray-200 text-gray-600 text-xs -ml-2 border-2 border-white">
-                                            +{hiddenCount}
-                                          </span>
-                                        )}
-                                      </>
-                                    );
-                                  })()}
-                                </div>
-                              )}
-                            </div>
-                            <div className="break-words whitespace-normal flex-1">
-                              {task.content}
-                            </div>
-                          </div>
-                        </TableCell>
-                        <TableCell className="text-center hidden md:table-cell">{task.count}</TableCell>
-                        <TableCell className="text-red-500 hidden md:table-cell">{task.deadline}</TableCell>
-                        <TableCell className="hidden md:table-cell">{task.createdAt}</TableCell>
-                      </TableRow>
-                    ))}
-                    </TableBody>
-                  </Table>
+                  <TableTask tasks={tasks} />
                 )}
               </div>
 
@@ -255,7 +179,15 @@ export default function TaskManagement() {
               </div>
             </TabsContent>
             <TabsContent value="completed">
-              <div className="p-8 text-center text-gray-500">Không có công việc đã hoàn thành</div>
+              <div className="overflow-x-auto">
+                {tasksLoading ? (
+                  <div className="flex justify-center items-center h-full">
+                    <LoadingSpinner />
+                  </div>
+                ) : (
+                  <TableTask tasks={taskCompleted} />
+                )}
+              </div>
             </TabsContent>
           </Tabs>
         </div>
